Some refactors:
- Update unit tests to support the use cases initialization after add koin;
- Create ReportDtoExt and testing the toMarkdownStatus method
