+++ popcornguineapigplugin/src/test/kotlin/com/github/codandotv/popcorn/domain/usecases/CheckArchitectureUseCaseTest.kt
@@ -16,7 +16,7 @@ class CheckArchitectureUseCaseTest {
 
     @Before
     fun before() {
-        checkArchitectureUseCase = CheckArchitectureUseCase()
+        checkArchitectureUseCase = CheckArchitectureUseCaseImpl()
     }
 
     @Test
