+++ popcornguineapigplugin/src/test/kotlin/com/github/codandotv/popcorn/domain/usecases/GetRightConfigurationNameUseCaseTest.kt
@@ -11,7 +11,7 @@ class GetRightConfigurationNameUseCaseTest {
 
     @Before
     fun before() {
-        getRightConfigurationNameUseCase = GetRightConfigurationNameUseCase()
+        getRightConfigurationNameUseCase = GetRightConfigurationNameUseCaseImpl()
     }
 
     @Test
